requirements.txt
py-cord
fuzzywuzzy
python-Levenshtein
python-dotenv
requests
python-dateutil